Renamed faucet to v1. As this wont work for our purposes but I am going to use it as a references as I build a new module.
