+++ x/faucetv1/internal/types/key.go
