+++ x/faucetv1/profile.go
@@ -1,6 +1,6 @@
 // +build !testnet
 
-package faucet
+package faucetv1
 
 // Faucet should not be added in mainnet version
 const profile = MAINNET
